python training project/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
openai==1.12.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==8.0.0
ollama==0.1.6

